requirements.txt
discord.py>=2.4      # poll + scheduled-event support
PyYAML>=6.0
python-dateutil>=2.8
